CalculatorService/requirements.txt
Werkzeug==2.3.3
Flask==2.3.2
pytest==7.2.0
requests
